refactor(DevSettings): split into file chunks
