+++ src/components/DevSettings/Separator.tsx
@@ -1,0 +1,3 @@
+export default function Separator() {
+  return <hr className="w-full border-slate-800 dark:border-slate-600" />;
+}
